test(gateway): split OpenAI model picker coverage
