+++ src/gateway/server-methods/models-list-result.openai-routes.test.ts
@@ -1,76 +1,19 @@
 import { describe, expect, it, vi } from "vitest";
 import type { ModelCatalogEntry, ModelCatalogSnapshot } from "../../agents/model-catalog.types.js";
-import type { createOpenAIModelRoutesResolver } from "../../agents/openai-model-routes.js";
-import { migrateLegacyConfig } from "../../commands/doctor/shared/legacy-config-migrate.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { withEnvAsync } from "../../test-utils/env.js";
 import { withOpenClawTestState } from "../../test-utils/openclaw-test-state.js";
 import { buildModelsListResult } from "./models-list-result.js";
+import {
+  catalogEntry,
+  listModels,
+  providerCatalogEntry,
+  WITHOUT_OPENAI_ENV_AUTH,
+} from "./models-list-result.openai-routes.test-support.js";
 import type { GatewayRequestContext } from "./types.js";
 
-const WITHOUT_OPENAI_ENV_AUTH = {
-  CODEX_API_KEY: undefined,
-  CODEX_HOME: "/__openclaw_models_list_test__/codex",
-  OPENAI_API_KEY: undefined,
-  OPENAI_BASE_URL: undefined,
-  OPENAI_OAUTH_TOKEN: undefined,
-  CHATGPT_OAUTH_TOKEN: undefined,
-} as const;
 const IMPLICIT_CODEX_RUNTIME = { id: "codex", source: "implicit" } as const;
 const IMPLICIT_OPENCLAW_RUNTIME = { id: "openclaw", source: "implicit" } as const;
-
-function catalogEntry(id: string, api: ModelCatalogEntry["api"]): ModelCatalogEntry {
-  return { id, name: id, provider: "openai", api };
-}
-
-function providerCatalogEntry(provider: string, id: string): ModelCatalogEntry {
-  return { ...catalogEntry(id, "openai-completions"), provider };
-}
-
-async function listModels(params: {
-  catalog: ModelCatalogEntry[];
-  cfg?: OpenClawConfig;
-  discoveryModes?: Record<string, "refreshable" | "runtime" | "static">;
-  routeResolverFactory?: typeof createOpenAIModelRoutesResolver;
-  view?: "all" | "configured" | "provider-config" | "default";
-}) {
-  const config = params.cfg ?? ({} as OpenClawConfig);
-  const context = {
-    getRuntimeConfig: () => config,
-    loadGatewayModelCatalog: vi.fn(() => Promise.resolve(params.catalog)),
-    loadGatewayModelCatalogSnapshot: vi.fn(() =>
-      Promise.resolve({
-        agentId: "main",
-        agentDir: "/tmp/models-list-openai-agent",
-        config,
-        entries: params.catalog,
-        routeVariants: params.catalog,
-      }),
-    ),
-    logGateway: { debug: vi.fn() },
-  } as unknown as GatewayRequestContext;
-  return await buildModelsListResult({
-    context,
-    params: { view: params.view ?? "all" },
-    ...(params.discoveryModes
-      ? {
-          preloadedCatalog: {
-            agentId: "main",
-            config,
-            snapshot: { entries: params.catalog, routeVariants: params.catalog },
-          },
-          catalogProjector: {
-            metadataSnapshot: {
-              plugins: [
-                { id: "test-provider", modelCatalog: { discovery: params.discoveryModes } },
-              ],
-            },
-          } as never,
-        }
-      : {}),
-    ...(params.routeResolverFactory ? { routeResolverFactory: params.routeResolverFactory } : {}),
-  });
-}
 
 describe("models.list OpenAI routes", () => {
   it("does not reuse a preloaded catalog owned by another agent", async () => {
@@ -593,37 +536,6 @@ describe("models.list OpenAI routes", () => {
       "gpt-5.4",
     ]);
     expect(result.models.every((entry) => !("providerOrder" in entry))).toBe(true);
-  });
-
-  it("does not expose a configured GPT-5.6 alias beside named variants after doctor normalization", async () => {
-    const staleConfig = {
-      agents: {
-        defaults: {
-          model: { primary: "openai/gpt-5.6" },
-          models: {
-            "openai/gpt-5.6": { alias: "GPT" },
-            "openai/gpt-5.6-sol": {},
-            "openai/gpt-5.6-terra": {},
-            "openai/gpt-5.6-luna": {},
-          },
-        },
-      },
-    } as OpenClawConfig;
-    const cfg = migrateLegacyConfig(staleConfig).config ?? staleConfig;
-    const catalog = [
-      { ...catalogEntry("gpt-5.6-sol", "openai-responses"), providerOrder: 0 },
-      { ...catalogEntry("gpt-5.6-terra", "openai-responses"), providerOrder: 1 },
-      { ...catalogEntry("gpt-5.6-luna", "openai-responses"), providerOrder: 2 },
-    ];
-
-    await withEnvAsync({ ...WITHOUT_OPENAI_ENV_AUTH, OPENAI_API_KEY: "test-key" }, async () => {
-      const result = await listModels({ catalog, cfg, view: "configured" });
-      expect(result.models.map((entry) => entry.id)).toEqual([
-        "gpt-5.6-sol",
-        "gpt-5.6-terra",
-        "gpt-5.6-luna",
-      ]);
-    });
   });
 
   it("keeps public metadata for a provider-canonical model-level Platform route", async () => {
